requirements.txt
alabaster==1.0.0
argparse-addons==0.12.0
astroid==3.3.11
attrs==25.3.0
babel==2.17.0
bitstruct==8.21.0
build==1.3.0
cantools==40.6.0
certifi==2025.8.3
cfgv==3.4.0
charset-normalizer==3.4.3
clang-format==21.1.1
click==8.3.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.3
coverage==7.10.7
crccheck==1.3.1
cycler==0.12.1
dill==0.4.0
diskcache==5.6.3
distlib==0.4.0
docopt==0.6.2
docutils==0.21.2
filelock==3.19.1
fonttools==4.60.0
gcovr==8.4
gitdb==4.0.12
GitPython==3.1.45
hatchling==1.27.0
identify==2.6.14
idna==3.10
imagesize==1.4.1
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
latexcodec==3.0.1
lauterbach-trace32-rcl==1.1.4
lxml==6.0.2
MarkupSafe==3.0.2
matplotlib==3.10.6
mccabe==0.7.0
mypy==1.18.2
mypy_extensions==1.1.0
nodeenv==1.9.1
numpy==2.3.3
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.2
pandas-stubs==2.3.2.250827
pathspec==0.12.1
pillow==11.3.0
pipdeptree==2.28.0
platformdirs==4.4.0
pre_commit==4.3.0
pyarrow==21.0.0
pybtex==0.25.1
pybtex-docutils==1.0.3
Pygments==2.19.2
pykwalify==1.8.0
pylint==3.3.8
pymodbus==3.4.0 # HIL
pyparsing==3.2.5
pyproject_hooks==1.2.0
pyright==1.1.405
pyrsistent==0.20.0
pyserial==3.5
python-can==4.6.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyVISA==1.15.0
pywin32==311; platform_system == "Windows"
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
roman-numerals-py==3.1.0
rpds-py==0.27.1
RsInstrument==1.102.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.14
ruff==0.13.2
SciencePlots==2.1.1
scipy==1.16.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinx-tabs==3.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-bibtex==2.6.5
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-drawio==0.0.17
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
textparser==0.24.0
tomlkit==0.13.3
types-colorama==0.4.15.20250801
types-jsonschema==4.25.1.20250822
types-pytz==2025.2.0.20250809
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
uptime==3.0.1
urllib3==2.5.0
virtualenv==20.34.0
wrapt==1.17.3
